add getTeamCountries(), move ftp test
add a convenience method Connection::getTeamCountries() and DB/Api implementations of it.

move FTP test to where it belongs.
Make the test use an associated ftp config instead of a hardcoded one.

Call FTP test from the plugin to unload if it fails
